:zap:Adding route ProductsbyCategory
