+++ src/Library/UseFetch.js
@@ -41,22 +41,22 @@ export function ProductsId(productId) {
         fetch(`${API}${productsEndPoint}${productId}`)
             .then((response) => response.json())
             .then((res) => setData(res));
+            console.log(data)
     }, [productId]);
 
     return data;
 }
-export function ProductsSearchbyCategory(inputCategoryId) {
-    const [data, setData] = useState([]);
-    useEffect(() => {
-        fetch(`${API}${productsEndPoint}`)
+export async function ProductsSearchbyCategory(inputCategoryId) {
+  inputCategoryId = parseInt(inputCategoryId)
+  return fetch(`${API}${productsEndPoint}`)
             .then((response) => response.json())
-            .then((res) => setData(res));
-    }, [inputCategoryId]);
-    let searchResults = [];
-    for (let i = 0; i < data.length; i++) {
-        if (data[i].categoryId === inputCategoryId) {
-            searchResults.push(data[i])
-        }
-    }
-    return searchResults;
+            .then((data) => {
+                let searchResults = [];
+                for (let i = 0; i < data.length -1; i++) {
+                if (data[i].categoryId === inputCategoryId) {
+                    searchResults.push(data[i])
+                    }
+                }
+            return searchResults;
+            });
 }
